Moved fixtures to main conf file
