+++ tests/test_post.py
@@ -4,412 +4,9 @@ import pytest
 
 from pyvista.plotting.renderer import CameraPosition
 
-from ansys.mapdl.core import examples
 from ansys.mapdl.core.post import (
     COMPONENT_STRESS_TYPE, PRINCIPAL_TYPE, STRESS_TYPES
 )
-
-@pytest.fixture(scope="module")
-def static_solve(mapdl):
-    mapdl.mute = True
-    mapdl.finish()
-    mapdl.clear()
-
-    # cylinder and mesh parameters
-    # torque = 100
-    radius = 2
-    h_tip = 2
-    height = 20
-    elemsize = 0.5
-    # pi = np.arccos(-1)
-    force = 100 / radius
-    pressure = force / (h_tip * 2 * np.pi * radius)
-
-    mapdl.prep7()
-    mapdl.et(1, 186)
-    mapdl.et(2, 154)
-    mapdl.r(1)
-    mapdl.r(2)
-
-    # Aluminum properties (or something)
-    mapdl.mp("ex", 1, 10e6)
-    mapdl.mp("nuxy", 1, 0.3)
-    mapdl.mp("dens", 1, 0.1 / 386.1)
-    mapdl.mp("dens", 2, 0)
-
-    # Simple cylinder
-    for i in range(4):
-        mapdl.cylind(radius, "", "", height, 90 * (i - 1), 90 * i)
-
-    mapdl.nummrg("kp")
-
-    # mesh cylinder
-    mapdl.lsel("s", "loc", "x", 0)
-    mapdl.lsel("r", "loc", "y", 0)
-    mapdl.lsel("r", "loc", "z", 0, height - h_tip)
-    mapdl.lesize("all", elemsize * 2)
-    mapdl.mshape(0)
-    mapdl.mshkey(1)
-
-    mapdl.esize(elemsize)
-    mapdl.allsel("all")
-    mapdl.vsweep("ALL")
-    mapdl.csys(1)
-    mapdl.asel("s", "loc", "z", "", height - h_tip + 0.0001)
-    mapdl.asel("r", "loc", "x", radius)
-    mapdl.local(11, 1)
-
-    mapdl.csys(0)
-
-    # mesh the surface with SURF154
-    mapdl.aatt(2, 2, 2, 11)
-    mapdl.amesh("all")
-    mapdl.prep7()
-
-    # plot elements
-    # mapdl.eplot()
-
-    # Apply tangential pressure
-    mapdl.esel("S", "TYPE", "", 2)
-    mapdl.sfe("all", 2, "pres", "", pressure)
-
-    # Constrain bottom of cylinder/rod
-    mapdl.asel("s", "loc", "z", 0)
-    mapdl.nsla("s", 1)
-    mapdl.d("all", "all")
-    mapdl.allsel()
-
-    # new solution
-    mapdl.run("/SOLU")
-    mapdl.antype("static", "new")
-    # mapdl.eqslv('pcg', 1e-8)
-    mapdl.solve()
-
-    # necessary for any prnsol printouts
-    mapdl.header("off", "off", "off", "off", "off", "off")
-    nsigfig = 10
-    mapdl.format("", "E", nsigfig + 9, nsigfig)
-    # mapdl.post1(mute=True)
-    # mapdl.set(1, 1)
-    mapdl.mute = False
-
-
-@pytest.fixture(scope="module")
-def plastic_solve(mapdl):
-    mapdl.mute = True
-    mapdl.finish()
-    mapdl.clear()
-    mapdl.input(examples.verif_files.vmfiles["vm273"])
-
-    mapdl.post1()
-    mapdl.set(1, 2)
-    mapdl.mute = False
-
-
-@pytest.fixture(scope="module")
-def contact_solve(mapdl):
-    mapdl.mute = True
-    mapdl.finish()
-    mapdl.clear()
-
-    # Based on tech demo 28.
-    mapdl.prep7()
-    # ***** Problem parameters ********
-    l = 76.2e-03/3     # Length of each plate,m
-    w = 31.75e-03/2    # Width of each plate,m
-    t = 3.18e-03     # Tickness of each plate,m
-    r1 = 7.62e-03    # Shoulder radius of tool,m
-    h = 15.24e-03    # Height of tool, m
-    l1 = r1          # Starting location of tool on weldline
-    l2 = l-l1
-    tcc1 = 2e06      # Thermal contact conductance b/w plates,W/m^2'C
-    tcc2 = 10        # Thermal contact conductance b/w tool &
-    # workpiece,W/m^2'C
-    fwgt = 0.95      # weight factor for distribution of heat b/w tool
-    # & workpiece
-    fplw = 0.8       # Fraction of plastic work converted to heat
-    uz1 = t/4000     # Depth of penetration,m
-    # ==========================================================
-    # * Material properties
-    # ==========================================================
-    # * Material properties for 304l stainless steel Plates
-    mapdl.mp("ex", 1, 193e9)  # Elastic modulus (N/m^2)
-    mapdl.mp("nuxy", 1, 0.3)  # Poisson's ratio
-    mapdl.mp("alpx", 1, 1.875e-5)  # Coefficient of thermal expansion, µm/m'c
-    # Fraction of plastic work converted to heat, 80%
-    mapdl.mp("qrate", 1, fplw)
-
-    # *BISO material model
-    EX = 193e9
-    ET = 2.8e9
-    EP = EX*ET/(EX-ET)
-    mapdl.tb("plas", 1, 1, "", "biso")  # Bilinear isotropic material
-    mapdl.tbdata(1, 290e6, EP)  # Yield stress & plastic tangent modulus
-    mapdl.mptemp(1, 0, 200, 400, 600, 800, 1000)
-    mapdl.mpdata("kxx", 1, 1, 16, 19, 21, 24, 29, 30)  # therm cond.(W/m'C)
-    mapdl.mpdata("c", 1, 1, 500, 540, 560, 590, 600, 610)  # spec heat(J/kg'C)
-    mapdl.mpdata("dens", 1, 1, 7894, 7744, 7631, 7518, 7406, 7406)  # kg/m^3
-
-    # * Material properties for PCBN tool
-    mapdl.mp("ex", 2, 680e9)  # Elastic modulus (N/m^2)
-    mapdl.mp("nuxy", 2, 0.22)  # Poisson's ratio
-    mapdl.mp("kxx", 2, 100)  # Thermal conductivity(W/m'C)
-    mapdl.mp("c", 2, 750)  # Specific heat(J/kg'C)
-    mapdl.mp("dens", 2, 4280)  # Density,kg/m^3
-
-    # ==========================================================
-    # * Geometry
-    # ==========================================================
-    # * Node for pilot node
-    mapdl.n(1, 0, 0, h)
-    # * Workpiece geometry (two rectangular plates)
-    mapdl.block(0, w, -l1, l2, 0, -t)
-    mapdl.block(0, -w, -l1, l2, 0, -t)
-    # * Tool geometry
-    mapdl.cyl4(0, 0, r1, 0, r1, 90, h)
-    mapdl.cyl4(0, 0, r1, 90, r1, 180, h)
-    mapdl.cyl4(0, 0, r1, 180, r1, 270, h)
-    mapdl.cyl4(0, 0, r1, 270, r1, 360, h)
-    mapdl.vglue(3, 4, 5, 6)
-
-    # ==========================================================
-    # * Meshing
-    # ==========================================================
-    mapdl.et(1, "SOLID226", 11)  # Coupled-field solid element,KEYOPT(1) is
-    # set to 11 for a structural-thermal analysis
-    mapdl.allsel()
-    ndiv1 = 2
-    ndiv2 = 5
-    ndiv3 = 1
-
-    mapdl.lsel("s", "", "", 4, 5)
-    mapdl.lsel("a", "", "", 14, 19, 5)
-    mapdl.lesize("all", "", "", ndiv1)
-    mapdl.lsel("s", "", "", 16, 17)
-    mapdl.lsel("a", "", "", 2, 7, 5)
-    mapdl.lesize("all", "", "", ndiv1)
-    mapdl.lsel("s", "", "", 1)
-    mapdl.lsel("a", "", "", 3)
-    mapdl.lsel("a", "", "", 6)
-    mapdl.lsel("a", "", "", 8)
-    mapdl.lsel("a", "", "", 13)
-    mapdl.lsel("a", "", "", 15)
-    mapdl.lsel("a", "", "", 18)
-    mapdl.lsel("a", "", "", 20)
-    mapdl.lesize("all", "", "", ndiv2)
-    mapdl.lsel("s", "", "", 9, "")
-    mapdl.lsel("a", "", "", 22)
-    mapdl.lesize("all", "", "", ndiv3)
-    mapdl.allsel("all")
-    mapdl.mshmid(2)  # midside nodes dropped
-    mapdl.vsweep(1)
-    mapdl.vsweep(2)
-    mapdl.vsel("u", "volume", "", 1, 2)
-    mapdl.mat(2)
-    mapdl.esize(0.005)
-    mapdl.numstr("NODE", 1000)
-    mapdl.vsweep("all")
-    mapdl.allsel("all")
-
-    # mapdl.eplot()
-    # ==========================================================
-    # * Contact Pairs
-    # ==========================================================
-    # * Define Rigid Surface Constraint on tool top surface
-    mapdl.et(2, "TARGE170")
-    mapdl.keyopt(2, 2, 1)  # User defined boundary condition on rigid
-    # target nodes
-
-    mapdl.et(3, "CONTA174")
-    mapdl.keyopt(3, 1, 1)  # To include Temp DOF
-    mapdl.keyopt(3, 2, 2)  # To include MPC contact algorithm
-    mapdl.keyopt(3, 4, 2)  # For a rigid surface constraint
-    mapdl.keyopt(3, 12, 5)  # To set the behavior of contact surface as a
-    # bonded (always)
-
-    mapdl.vsel("u", "volume", "", 1, 2)  # Selecting Tool volume
-    mapdl.allsel("below", "volume")
-    mapdl.nsel("r", "loc", "z", h)  # Selecting nodes on the tool top surface
-    mapdl.type(3)
-    mapdl.r(3)
-    mapdl.real(3)
-    mapdl.esln()
-    mapdl.esurf()  # Create contact elements
-    mapdl.allsel("all")
-
-    # * Define pilot node at the top of the tool
-    mapdl.nsel("s", "node", "", 1)
-    mapdl.tshap("pilo")
-    mapdl.type(2)
-    mapdl.real(3)
-    mapdl.e(1)  # Create target element on pilot node
-    mapdl.allsel()
-
-    # * Define contact pair between two plates
-    mapdl.et(6, "TARGE170")
-    mapdl.et(7, "CONTA174")
-    mapdl.keyopt(7, 1, 1)  # Displacement & Temp dof
-    mapdl.keyopt(7, 4, 3)  # To include Surface projection based method
-    mapdl.mat(1)
-    mapdl.asel("s", "", "", 5)
-    mapdl.nsla("", 1)
-    # mapdl.nplot()
-    mapdl.cm("tn.cnt", "node")  # Creating component on weld side of plate1
-
-    mapdl.asel("s", "", "", 12)
-    mapdl.nsla("", 1)
-    # mapdl.nplot()
-    mapdl.cm("tn.tgt", "node")  # Creating component on weld side of plate2
-
-    mapdl.allsel("all")
-    mapdl.type(6)
-    mapdl.r(6)
-    mapdl.rmodif(6, 14, tcc1)  # A real constant TCC,Thermal contact
-    # conductance coeffi. b/w the plates, W/m^2'C
-    mapdl.rmodif(6, 35, 1000)  # A real constant TBND,Bonding temperature
-    # for welding, 'C
-    mapdl.real(6)
-    mapdl.cmsel("s", "tn.cnt")
-    # mapdl.nplot()
-    mapdl.esurf()
-    mapdl.type(7)
-    mapdl.real(6)
-    mapdl.cmsel("s", "tn.tgt")
-    mapdl.esurf()
-    mapdl.allsel("all")
-
-    # * Define contact pair between tool & workpiece
-    mapdl.et(4, "TARGE170")
-    mapdl.et(5, "CONTA174")
-    mapdl.keyopt(5, 1, 1)  # Displacement & Temp dof
-    mapdl.keyopt(5, 5, 3)  # Close gap/reduce penetration with auto cnof
-    mapdl.keyopt(5, 9, 1)  # Exclude both initial penetration or gap
-    mapdl.keyopt(5, 10, 0)  # Contact stiffness update each iteration
-    # based
-
-    # Bottom & lateral(all except top) surfaces of tool for target
-    mapdl.vsel("u", "volume", "", 1, 2)
-    mapdl.allsel("below", "volume")
-    mapdl.nsel("r", "loc", "z", 0, h)
-    mapdl.nsel("u", "loc", "z", h)
-    mapdl.type(4)
-    mapdl.r(5)
-    mapdl.tb("fric", 5, 6)  # Definition of friction co efficient at
-    # different temp
-    mapdl.tbtemp(25)
-    mapdl.tbdata(1, 0.4)  # friction co-efficient at temp 25
-    mapdl.tbtemp(200)
-    mapdl.tbdata(1, 0.4)  # friction co-efficient at temp 200
-    mapdl.tbtemp(400)
-    mapdl.tbdata(1, 0.4)  # friction co-efficient at temp 400
-    mapdl.tbtemp(600)
-    mapdl.tbdata(1, 0.3)  # friction co-efficient at temp 600
-    mapdl.tbtemp(800)
-    mapdl.tbdata(1, 0.3)  # friction co-efficient at temp 800
-    mapdl.tbtemp(1000)
-    mapdl.tbdata(1, 0.2)  # friction co-efficient at temp 1000
-    mapdl.rmodif(5, 9, 500e6)  # Max.friction stress
-    mapdl.rmodif(5, 14, tcc2)  # Thermal contact conductance b/w tool and
-    # workpiece, 10 W/m^2'C
-    mapdl.rmodif(5, 15, 1)  # A real constant FHTG,the fraction of
-    # frictional dissipated energy converted
-    # into heat
-    mapdl.rmodif(5, 18, fwgt)  # A real constant  FWGT, weight factor for
-    # the distribution of heat between the
-    # contact and target surfaces, 0.95
-    mapdl.real(5)
-    mapdl.mat(5)
-    mapdl.esln()
-    mapdl.esurf()
-    mapdl.allsel("all")
-
-    # Top surfaces of plates nodes for contact
-    mapdl.vsel("s", "volume", "", 1, 2)
-    mapdl.allsel("below", "volume")
-    mapdl.nsel("r", "loc", "z", 0)
-    mapdl.type(5)
-    mapdl.real(5)
-    mapdl.esln()
-    mapdl.esurf()
-    mapdl.allsel("all")
-
-    # ==========================================================
-    # * Boundary conditions
-    # ==========================================================
-    mapdl.tref(25)  # Reference temperature 25'C
-    mapdl.allsel()
-    mapdl.nsel("all")
-    mapdl.ic("all", "temp", 25)  # Initial condition at nodes,temp 25'C
-
-    # Mechanical Boundary Conditions
-    # 20% ends of the each plate is constraint
-    mapdl.nsel("s", "loc", "x", -0.8*w, -w)
-    mapdl.nsel("a", "loc", "x", 0.8*w, w)
-    mapdl.d("all", "uz", 0)  # Displacement constraint in x-direction
-    mapdl.d("all", "uy", 0)  # Displacement constraint in y-direction
-    mapdl.d("all", "ux", 0)  # Displacement constraint in z-direction
-    mapdl.allsel("all")
-
-    # Bottom of workpiece is constraint in z-direction
-    mapdl.nsel("s", "loc", "z", -t)
-    mapdl.d("all", "uz")  # Displacement constraint in z-direction
-    mapdl.allsel("all")
-
-    # Thermal Boundary Conditions
-    # Convection heat loss from the workpiece surfaces
-    mapdl.vsel("s", "volume", "", 1, 2)  # Selecting the workpiece
-    mapdl.allsel("below", "volume")
-    mapdl.nsel("r", "loc", "z", 0)
-    mapdl.nsel("a", "loc", "x", -w)
-    mapdl.nsel("a", "loc", "x", w)
-    mapdl.nsel("a", "loc", "y", -l1)
-    mapdl.nsel("a", "loc", "y", l2)
-    mapdl.sf("all", "conv", 30, 25)
-
-    # Convection (high)heat loss from the workpiece bottom
-    mapdl.nsel("s", "loc", "z", -t)
-    mapdl.sf("all", "conv", 300, 25)
-    mapdl.allsel("all")
-
-    # Convection heat loss from the tool surfaces
-    mapdl.vsel("u", "volume", "", 1, 2)  # Selecting the tool
-    mapdl.allsel("below", "volume")
-    mapdl.csys(1)
-    mapdl.nsel("r", "loc", "x", r1)
-    mapdl.nsel("a", "loc", "z", h)
-    mapdl.sf("all", "conv", 30, 25)
-    mapdl.allsel("all")
-
-    # Constraining all DOFs at pilot node except the Temp DOF
-    mapdl.d(1, "all")
-    mapdl.ddele(1, "temp")
-    mapdl.allsel("all")
-
-    # mapdl.eplot()
-    # ==========================================================
-    # * Solution
-    # ==========================================================
-    mapdl.run('/solu')
-    mapdl.antype(4)  # Transient analysis
-    mapdl.lnsrch('on')
-    mapdl.cutcontrol('plslimit', 0.15)
-    mapdl.kbc(0)  # Ramped loading within a load step
-    mapdl.nlgeom("on")  # Turn on large deformation effects
-    mapdl.timint("off", "struc")  # Structural dynamic effects are turned off.
-    mapdl.nropt('unsym')
-
-    # Load Step1
-    mapdl.time(1)
-    mapdl.nsubst(5, 10, 2)
-    mapdl.d(1, "uz", -uz1)  # Tool plunges into the workpiece
-    mapdl.outres("all", "all")
-    mapdl.allsel()
-    mapdl.solve()
-
-    mapdl.post1()
-    mapdl.allsel()
-    mapdl.set('last')
-    mapdl.mute = False
 
 
 # must be run first before loading a result
